+++ lib/src/models/components/components.dart
@@ -32,6 +32,16 @@ class CalendarComponents<T extends Object?> {
   /// Styles used by the schedule view.
   final ScheduleComponentStyles? scheduleComponentStyles;
 
+  /// Default override for the overlay widgets.
+  ///
+  /// If a more specific builder is provided in [multiDayComponents] or [monthComponents], that will be used instead.
+  final OverlayBuilders<T>? overlayBuilders;
+
+  /// Default styles for the overlay widgets.
+  ///
+  /// If another style is provided in [multiDayComponentStyles] or [monthComponentStyles], that will be used instead.
+  final OverlayStyles? overlayStyles;
+
   CalendarComponents({
     this.monthComponents,
     this.monthComponentStyles,
@@ -39,6 +49,8 @@ class CalendarComponents<T extends Object?> {
     this.multiDayComponentStyles,
     this.scheduleComponents,
     this.scheduleComponentStyles,
+    this.overlayBuilders,
+    this.overlayStyles,
   });
 }